requirements.txt
amqp==5.0.6
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.4.0
braintree==4.7.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==4.4.2
Django==3.1.7
django-elasticsearch-dsl==7.2.0
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
elasticsearch==7.14.1
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
kombu==5.0.2
kwargs==1.0.1
MarkupSafe==1.1.1
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.17
psycopg2==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.17.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
sentry-sdk==1.0.0
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
